try to delete emssages just if they are older than 48 hours

+++ plugins/applications/private/any_message.py
@@ -37,8 +37,13 @@ async def on_test_delhistory(update: Update, context: ContextTypes.DEFAULT_TYPE,
     rabbit_file_id = "AgACAgQAAxkBAAIF4WRCV9_H-H1tQHnA2443fXtcVy4iAAKkujEbkmDgUYIhRK-rWlZHAQADAgADeAADLwQ"
     sent_message = await update.message.reply_photo(rabbit_file_id)
 
+    now = utilities.now()
+
     messages: List[PrivateChatMessage] = private_chat_messages.get_messages(session, update.effective_user.id)
     for message in messages:
+        if not message.can_be_deleted(now):
+            continue
+
         logger.debug(f"deleting message {message.message_id} from chat {update.effective_user.id}")
         await context.bot.delete_message(update.effective_user.id, message.message_id)
         message.set_revoked(reason="/delhistory command")
